+++ src/Events/HandlingWriteRequestEvent.php
@@ -13,6 +13,7 @@
 
 namespace IceHawk\IceHawk\Events;
 
+use IceHawk\IceHawk\Interfaces\ProvidesCookieData;
 use IceHawk\IceHawk\Interfaces\ProvidesRequestInfo;
 use IceHawk\IceHawk\Interfaces\ProvidesWriteRequestData;
 use IceHawk\IceHawk\Interfaces\ProvidesWriteRequestInputData;
@@ -41,5 +42,10 @@ final class HandlingWriteRequestEvent implements CarriesEventData
 	public function getRequestInput() : ProvidesWriteRequestInputData
 	{
 		return $this->request->getInput();
+	}
+
+	public function getRequestCookies() : ProvidesCookieData
+	{
+		return $this->request->getCookies();
 	}
 }