+++ tests/Unit/Defaults/FinalReadRequestResponderTest.php
@@ -13,6 +13,7 @@
 
 namespace IceHawk\IceHawk\Tests\Unit\Defaults;
 
+use IceHawk\IceHawk\Defaults\Cookies;
 use IceHawk\IceHawk\Defaults\FinalReadResponder;
 use IceHawk\IceHawk\Defaults\RequestInfo;
 use IceHawk\IceHawk\Exceptions\UnresolvedRequest;
@@ -36,7 +37,7 @@ class FinalReadRequestResponderTest extends \PHPUnit_Framework_TestCase
 		);
 
 		$routeRequest = new RouteRequest( $requestInfo->getUri(), $requestInfo->getMethod() );
-		$requestData  = new ReadRequest( $requestInfo, new ReadRequestInput( [] ) );
+		$requestData = new ReadRequest( $requestInfo, new Cookies( [] ), new ReadRequestInput( [] ) );
 
 		try
 		{